requirements.txt
notion-client==0.6.0
google-api-python-client==2.23.0
google-auth-oauthlib==0.4.6
pyyaml==5.4.1
pandas==1.3.3
pytest~=6.2.4